requirements.txt
streamlit
basicsr
realesrgan
numpy
pillow
torch
torchvision
